todolist/pyproject.toml
[tool.poetry]
name = "ToDo List App"
version = "0.1.0"
description = "To do application based on Django 4.1.3"
authors = ["Sergey Protsenko"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.3"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.0.0"
gunicorn = "^20.1.0"
django-filter = "^22.1"
marshmallow = "^3.18.0"
marshmallow-dataclass = "^8.5.9"
pytest = "^7.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
